TP update_data handles multiple stores
