+++ pootle/apps/pootle_data/getters.py
@@ -18,8 +18,8 @@ from .directory_data import DirectoryDataTool
 from .language_data import LanguageDataTool
 from .project_data import (
     ProjectDataTool, ProjectResourceDataTool, ProjectSetDataTool)
-from .store_data import StoreDataTool, StoreDataUpdater, StoreUpdater
-from .tp_data import TPDataTool, TPDataUpdater
+from .store_data import StoreDataTool, StoreDataUpdater
+from .tp_data import TPDataTool, TPDataUpdater, TPUpdater
 
 
 @getter(data_tool, sender=Store)
@@ -27,9 +27,9 @@ def store_data_tool_getter(**kwargs_):
     return StoreDataTool
 
 
-@getter(data_updater, sender=Store)
+@getter(data_updater, sender=TranslationProject)
 def store_updater_getter(**kwargs_):
-    return StoreUpdater
+    return TPUpdater
 
 
 @getter(data_updater, sender=StoreDataTool)
