+++ pootle/apps/pootle_data/tp_data.py
@@ -9,8 +9,11 @@
 from django.db.models import Max, Sum
 from django.db.models.functions import Coalesce
 
+from pootle.core.contextmanagers import keep_data
 from pootle.core.decorators import persistent_property
 from pootle.core.delegate import revision
+from pootle.core.signals import update_revisions, update_data
+from pootle_app.models import Directory
 from pootle_data.models import StoreChecksData, StoreData
 
 from .utils import DataUpdater, RelatedStoresDataTool
@@ -151,3 +154,22 @@ class TPDataTool(RelatedStoresDataTool):
     @property
     def max_unit_revision(self):
         return self.context.data.max_unit_revision
+
+
+class TPUpdater(object):
+
+    def __init__(self, tp, object_list):
+        self.tp = tp
+        self.object_list = object_list
+
+    def update(self):
+        dirs = set()
+        with keep_data(suppress=(self.tp.__class__, )):
+            with keep_data(signals=(update_revisions, )):
+                for store in self.object_list:
+                    update_data.send(store.__class__, instance=store)
+                    dirs.add(store.parent_id)
+        if dirs:
+            update_revisions.send(
+                Directory,
+                object_list=Directory.objects.filter(id__in=dirs))
